+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/App.kt
@@ -5,11 +5,13 @@ import androidx.compose.runtime.*
 import com.altmann.choresmanager.ui.screens.calendar.CalendarScreen
 import com.altmann.choresmanager.ui.screens.calendar.CalendarViewModel
 import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
 @Preview
 fun App() {
     MaterialTheme {
-        CalendarScreen(CalendarViewModel())
+        val viewModel = koinViewModel<CalendarViewModel>()
+        CalendarScreen(viewModel)
     }
 }