+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/di/KoinModule.kt
@@ -1,0 +1,30 @@
+package com.altmann.choresmanager.di
+
+import com.altmann.choresmanager.database.UserDao
+import com.altmann.choresmanager.local.DatabaseDriverFactory
+import com.altmann.choresmanager.local.LocalDatabase
+import com.altmann.choresmanager.ui.screens.calendar.CalendarViewModel
+import org.koin.core.KoinApplication
+import org.koin.core.context.startKoin
+import org.koin.core.module.Module
+import org.koin.core.module.dsl.viewModelOf
+import org.koin.dsl.module
+
+
+expect val targetModule: Module
+
+val sharedModule = module {
+    single<UserDao> {
+        UserDao(get())
+    }
+    single<LocalDatabase> { LocalDatabase(get()) }
+    viewModelOf(::CalendarViewModel)
+
+}
+
+fun initializeKoin(config: (KoinApplication.() -> Unit)? = null) {
+    startKoin {
+        modules(targetModule, sharedModule)
+        config?.invoke(this)
+    }
+}