+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/User.kt
@@ -1,0 +1,6 @@
+package com.altmann.choresmanager.models.chores
+
+data class User(
+    val userId: Int,
+    val name: String
+)