+++ composeApp/src/jvmMain/kotlin/com/altmann/choresmanager/main.kt
@@ -4,6 +4,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.WindowState
 import androidx.compose.ui.window.application
+import com.altmann.choresmanager.di.initializeKoin
 
 fun main() = application {
     Window(
@@ -11,6 +12,7 @@ fun main() = application {
         title = "ChoresManager",
         state = WindowState(width = 800.dp, height = 820.dp)
     ) {
+        initializeKoin()
         App()
     }
 }